move filelist
