+++ build-system/app-index/template.js
@@ -20,35 +20,12 @@
 'use strict';
 
 const boilerPlate = require('./boilerplate');
-const documentModes = require('./document-modes');
 const headerLinks = require('./header-links');
 const ProxyForm = require('./proxy-form');
-const {
-  AmpState,
-  addRequiredExtensionsToHead,
-  containsExpr,
-} = require('./amphtml-helpers');
-const {appendQueryParamsToUrl, replaceLeadingSlash} = require('./url');
+const {addRequiredExtensionsToHead} = require('./amphtml-helpers');
+const {FileList} = require('./file-list');
 const {html, joinFragments} = require('./html');
-const {KeyValueOptions} = require('./form');
 const {SettingsModal, SettingsOpenButton} = require('./settings');
-
-const examplesPathRegex = /^\/examples\//;
-const htmlDocRegex = /\.html$/;
-
-const fileListEndpoint = query =>
-  appendQueryParamsToUrl('/dashboard/api/listing', query);
-
-const ampStateKey = (...keys) => keys.join('.');
-
-const selectModeStateId = 'documentMode';
-const selectModeStateKey = 'selectModePrefix';
-const selectModeKey = ampStateKey(selectModeStateId, selectModeStateKey);
-
-const fileListEndpointStateId = 'fileListEndpoint';
-const fileListEndpointStateKey = 'src';
-const fileListEndpointKey = ampStateKey(
-    fileListEndpointStateId, fileListEndpointStateKey);
 
 
 const HeaderLink = ({name, href, divider}) => html`
@@ -76,119 +53,12 @@ const Header = ({isMainPage, links}) => html`
     </ul>
   </header>`;
 
-const FileListSearchInput = ({basepath}) => html`
-  <input type="text"
-    class="file-list-search"
-    placeholder="Fuzzy Search"
-    pattern="[a-zA-Z0-9-]+"
-    on="input-debounced: AMP.setState({
-      ${fileListEndpointStateId}: {
-        ${fileListEndpointStateKey}: '${fileListEndpoint({
-          path: basepath,
-          search: '',
-        })}' + event.value
-      }
-    })">`;
 
 const HeaderBackToMainLink = () => html`<a href="/">← Back to main</a>`;
 
-const ExamplesDocumentModeSelect = () => html`
-  <label for="examples-mode-select">
-    Document mode:
-    <select id="examples-mode-select"
-        on="change:AMP.setState({
-          ${selectModeStateId}: {
-            ${selectModeStateKey}: event.value
-          }
-        })">
-      ${KeyValueOptions(documentModes)}
-    </select>
-  </label>`;
-
-
-const linksToExample = (shouldContainBasepath, opt_name) =>
-  examplesPathRegex.test(shouldContainBasepath) &&
-    htmlDocRegex.test(opt_name || shouldContainBasepath);
-
-
-const ExamplesSelectModeOptional = ({basepath, selectModePrefix}) =>
-  !examplesPathRegex.test(basepath + '/') ? '' : ExamplesDocumentModeSelect({
-    selectModePrefix,
-  });
-
-
-const FileListItem = ({name, href, boundHref}) =>
-  html`<div class="file-link-container" role=listitem>
-    <a class="file-link"
-      ${boundHref ? `[href]="${boundHref}" ` : ''}
-      ${href ? `href="${href}" ` : ''}>
-      ${name}
-    </a>
-  </div>`;
-
-
-const PlaceholderFileListItem = ({name, href, selectModePrefix}) =>
-  linksToExample(href) ?
-    FileListItem({
-      name,
-      href: selectModePrefix + replaceLeadingSlash(href, ''),
-      boundHref: `(${selectModeKey} || '${selectModePrefix}') + '${
-        replaceLeadingSlash(href, '')}'`,
-    }) :
-    FileListItem({href, name});
-
-
-const maybePrefixExampleDocHref = (basepath, name, selectModePrefix) =>
-  (linksToExample(basepath, name) ?
-    replaceLeadingSlash(basepath, selectModePrefix) :
-    basepath) +
-  name;
-
-const FileList = ({basepath, fileSet, selectModePrefix}) => joinFragments([
-  AmpState(fileListEndpointStateId, {
-    [fileListEndpointStateKey]: fileListEndpoint({path: basepath}),
-  }),
-
-  html`<amp-list [src]="${fileListEndpointKey}"
-    src="${fileListEndpoint({path: basepath})}"
-    items="."
-    layout="fixed-height"
-    width="auto"
-    height="568px"
-    class="file-list custom-loader">
-
-    <div fallback>Failed to load data.</div>
-
-    <div placeholder>
-      <div role=list>
-        ${joinFragments(fileSet, ({name, href}) =>
-          PlaceholderFileListItem({name, href, selectModePrefix}))}
-      </div>
-    </div>
-
-    <template type="amp-mustache">
-      ${FileListItem({
-        href: `${basepath}{{.}}`,
-        boundHref: containsExpr(
-            '\'{{.}}\'',
-            '\'.html\'',
-            `(${selectModeKey} || '${selectModePrefix}') +` +
-                `'${replaceLeadingSlash(basepath, '')}{{.}}'`,
-            `'${basepath}{{.}}'`),
-        name: '{{.}}',
-      })}
-    </template>
-
-    <div overflow
-      role="button"
-      aria-label="Show more"
-      class="list-overflow">
-      Show more
-    </div>
-  </amp-list>`,
-]);
 
 const ProxyFormOptional = ({isMainPage}) => isMainPage ? ProxyForm() : '';
+
 
 const renderTemplate = ({
   basepath,
@@ -217,31 +87,7 @@ const renderTemplate = ({
       ${Header({isMainPage, links: headerLinks})}
       ${ProxyFormOptional({isMainPage})}
     </div>
-    <div class="file-list-container">
-      <div class="wrap">
-        <div class="file-list-heading">
-          <h3 class="code" id="basepath">
-            ${basepath}
-          </h3>
-          ${FileListSearchInput({basepath})}
-          <div class="file-list-right-section">
-            ${AmpState(selectModeStateId, {
-              [selectModeStateKey]: selectModePrefix,
-            })}
-            ${ExamplesSelectModeOptional({basepath, selectModePrefix})}
-            <a href="/~" class="underlined">List root directory</a>
-          </div>
-        </div>
-        ${FileList({
-          basepath,
-          selectModePrefix,
-          fileSet: fileSet.map(name => ({
-            name,
-            href: maybePrefixExampleDocHref(basepath, name, selectModePrefix),
-          })),
-        })}
-      </div>
-    </div>
+    ${FileList({basepath, selectModePrefix, fileSet})}
     <div class="center">
       Built with 💙  by
       <a href="https://ampproject.org" class="underlined">the AMP Project</a>.
